+++ lib/tagged/constructor.ex
@@ -49,19 +49,26 @@ defmodule Tagged.Constructor do
   ##############################################################################
 
   @doc false
-  @spec __deftagged__(Keyword.t()) :: Macro.t()
+  @spec __deftagged__(Keyword.t()) :: [Macro.t()]
   def __deftagged__(params) do
     tag = Keyword.get(params, :tag)
     ex_tag = Keyword.get(params, :ex_tag)
     module = Keyword.get(params, :module)
     name = Keyword.get(params, :name)
     arity = Keyword.get(params, :arity)
+    p? = Keyword.get(params, :__private__)
 
-    gen_constructor(tag, module, name, arity, ex_tag)
+    [
+      gen_doc(ex_tag, module, name, arity, p?),
+      gen_constructor(tag, name, arity, p?)
+    ]
   end
 
-  defp gen_constructor(tag, module, name, 0, ex_tag) do
-    quote do
+  @spec gen_doc(String.t(), module(), atom(), integer(), boolean()) :: Macro.t()
+  defp gen_doc(_, _, _, _, true), do: nil
+
+  defp gen_doc(ex_tag, module, name, 0, false) do
+    quote location: :keep do
       @doc """
       Constructor `#{unquote(name)}/0` for `#{unquote(ex_tag)}` tags.
 
@@ -73,12 +80,10 @@ defmodule Tagged.Constructor do
           :not_#{unquote(ex_tag)}
 
       """
-      defmacro unquote(name)(), do: unquote(tag)
     end
   end
 
-  defp gen_constructor(tag, module, name, arity, ex_tag) do
-    cons_args = Macro.generate_arguments(arity, nil)
+  defp gen_doc(ex_tag, module, name, arity, false) do
     ex_match = for(_ <- 1..arity, do: "_") |> Enum.join(", ")
     ex_vals = for(i <- 1..arity, do: "#{i}") |> Enum.join(", ")
 
@@ -86,9 +91,9 @@ defmodule Tagged.Constructor do
     ex_hit = "{:#{ex_tag}, #{ex_vals}}"
     ex_miss = "{:not_#{ex_tag}, #{ex_vals}}"
 
-    quote do
+    quote location: :keep do
       @doc """
-      Constructor `#{unquote(name)}/#{unquote(arity)}` for `#{unquote(tag)}`
+      Constructor `#{unquote(name)}/#{unquote(arity)}` for `:#{unquote(ex_tag)}`
       tagged value tuples. Can also be used to destructure tuples.
 
           iex> require #{unquote(module)}
@@ -99,7 +104,40 @@ defmodule Tagged.Constructor do
           #{unquote(ex_miss)}
 
       """
+    end
+  end
+
+  @spec gen_constructor(atom(), atom(), integer(), boolean()) :: Macro.t()
+  defp gen_constructor(tag, name, 0, false) do
+    quote location: :keep do
+      defmacro unquote(name)(), do: unquote(tag)
+    end
+  end
+
+  defp gen_constructor(tag, name, 0, true) do
+    quote location: :keep do
+      defmacrop unquote(name)(), do: unquote(tag)
+    end
+  end
+
+  defp gen_constructor(tag, name, arity, false) do
+    cons_args = Macro.generate_arguments(arity, nil)
+
+    quote location: :keep do
       defmacro unquote(name)(unquote_splicing(cons_args)) do
+        args = [unquote_splicing(cons_args)]
+        tag = unquote(tag)
+
+        quote do: {unquote(tag), unquote_splicing(args)}
+      end
+    end
+  end
+
+  defp gen_constructor(tag, name, arity, true) do
+    cons_args = Macro.generate_arguments(arity, nil)
+
+    quote location: :keep do
+      defmacrop unquote(name)(unquote_splicing(cons_args)) do
         args = [unquote_splicing(cons_args)]
         tag = unquote(tag)