+++ lib/tagged/typedef.ex
@@ -51,52 +51,63 @@ defmodule Tagged.Typedef do
   ##
   ##############################################################################
 
-  @typep macro?() :: Macro.t() | false
-
   @doc false
-  @spec __deftagged__(Keyword.t()) :: macro?
+  @spec __deftagged__(Keyword.t()) :: Macro.t() | []
   def __deftagged__(params) do
     with true <- Keyword.get(params, :type, true),
          name = Keyword.get(params, :name),
          args = Keyword.get(params, :args),
          tag = Keyword.get(params, :tag),
+         p? = Keyword.get(params, :__private__),
          of_type = Keyword.get(params, :of, args) do
-      gen_typedef(name, tag, List.wrap(of_type))
+      gen_typedef(name, tag, List.wrap(of_type), p?)
     else
       _ -> []
     end
   end
 
   @doc false
-  @spec gen_typedef(atom(), atom(), Macro.t()) :: Macro.t()
-  def gen_typedef(name, tag, []) do
-    quote do
+  @spec gen_typedef(atom(), atom(), Macro.t(), boolean()) :: Macro.t()
+  def gen_typedef(name, tag, [], false) do
+    quote location: :keep do
       @type unquote(name)() :: unquote(tag)
     end
   end
 
-  def gen_typedef(name, tag, of_type) do
-    #
-    # FIXME: Generate proper arity type declarations when a type argument is
-    # `_`, to allow for parameterized type usage (as we did before).
-    #
-    # E.g.
-    #     deftagged ok(_)
-    #     => @type ok(t1) :: {:ok, t1}
-    #        @type ok() :: ok(term())
-    #
-    # FIXME: Multiple arity definitions of the same name will cause
-    # redefinitions of @type t/0 and must instead accumulate and declare the sum
-    # type in a __before_compile__ hook.
-    #
-    # E.g.
-    #     deftagged error(term())
-    #     => @type error() :: {:error, term()}
-    #     deftagged error()
-    #     => ** (CompileError) ... type error/0 is already defined
-    #
-    quote do
+  def gen_typedef(name, tag, [], true) do
+    quote location: :keep do
+      @typep unquote(name)() :: unquote(tag)
+    end
+  end
+
+  #
+  # FIXME: Generate proper arity type declarations when a type argument is
+  # `_`, to allow for parameterized type usage (as we did before).
+  #
+  # E.g.
+  #     deftagged ok(_)
+  #     => @type ok(t1) :: {:ok, t1}
+  #        @type ok() :: ok(term())
+  #
+  # FIXME: Multiple arity definitions of the same name will cause
+  # redefinitions of @type t/0 and must instead accumulate and declare the sum
+  # type in a __before_compile__ hook.
+  #
+  # E.g.
+  #     deftagged error(term())
+  #     => @type error() :: {:error, term()}
+  #     deftagged error()
+  #     => ** (CompileError) ... type error/0 is already defined
+  #
+  def gen_typedef(name, tag, of_type, false) do
+    quote location: :keep do
       @type unquote(name)() :: {unquote(tag), unquote_splicing(of_type)}
+    end
+  end
+
+  def gen_typedef(name, tag, of_type, true) do
+    quote location: :keep do
+      @typep unquote(name)() :: {unquote(tag), unquote_splicing(of_type)}
     end
   end
 end