+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+### Fixed
+
+- The `|> with_...()` macros didn't work unless the module was `import`:ed
+
+### Added
+
+- Definition of private tagged tuples.
+
 ## [v0.4.1] - 2020-05-31
 
 ### Fixed